Add files via upload
contains the netbeans package folders for Orb.  you can find the .java files in /src/ and the media files in /media/ if needed for creating the project in your ide.  all media files are referred to with "./media/file.fileextension" in the java code, so you'll want to keep that structure for the media.